refactor : Trap 리팩토링
ProcessActive에서 로직을 작성
Active로 ProcessActive를 호출하는 방식을 사용함
