--- Assets/Project/NoSeungJu/Script/Disposable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a2dfb0c7299d8954ba59a499abfc5e08
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
