+++ Assets/Project/Public/Scripts/Item/HpItem.cs
@@ -6,6 +6,7 @@ public class HpItem : Item
 {
     [Header("Èú·® (±âº»Èú·® 1)")]
     [SerializeField] int _healAmount = 1;
+
     protected override void Use(PlayerController player)
     {
         player.TakeHeal(_healAmount);
