+++ Assets/Project/Public/Scripts/Item/Item.cs
@@ -1,19 +1,23 @@
-using System.Collections;
-using System.Collections.Generic;
-using Unity.VisualScripting;
 using UnityEngine;
 
 public abstract class Item : Trap
 {
+    PlayerController player;
+
     protected override void OnTriggerEnter2D(Collider2D collision)
     {
-        base.OnTriggerEnter2D(collision);
-        if(collision.gameObject.tag == "Player")
+        if (collision.gameObject.tag == "Player")
         {
-            PlayerController player = collision.gameObject.GetComponent<PlayerController>();
-            Use(player);
+            player = collision.gameObject.GetComponent<PlayerController>();
+            Active();
             Delete();
         }
+    }
+
+    protected override void ProcessActive()
+    {
+        PlayerController playerController = player;
+        Use(playerController);
     }
 
     protected abstract void Use(PlayerController player);
